Add MIT License and README files
Two new documents are added in this commit - a License file stating project usage rights and a README file providing brief introduction to the project. The MIT license has been chosen to allow the consumers of the software to have flexibility in its usage. The README file simply describes the core technologies being demonstrated in the repo.