enhance: Update RabbitMQ configuration and make Book Serializable
Changes were made to update the configuration properties prefix to 'spring.rabbitmq' and change the port number to '5672'. Adjustments have also been made in Book class to implement Serializable in order to allow the object to be converted into a byte stream. A new class, RabbitMqConfig, was added to establish rabbitmq configurations. All these changes were made to improve adaptability and scalability of the app.

In 'application.yml', updated the RabbitMQ details so that configurations are up to date. This enhances the application's performance and consistency with the external rabbitmq service. Improvements to code and configuration ensure better application performance and robustness.